feed_ingestor/requirements.txt
Pillow
opencv-python
redis